// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        PlanModel: {
            /** Plan Id */
            plan_id?: string | null;
            /** Plan Name */
            plan_name: string;
            /** Suites Ids */
            suites_ids?: string | null;
            /** Suite Tags */
            suite_tags?: {
                [key: string]: unknown;
            } | null;
            /** Created At */
            created_at?: string | null;
            /** Updated At */
            updated_at?: string | null;
            /**
             * Status
             * @default active
             */
            status: string;
            /** Owner User Id */
            owner_user_id?: string | null;
            /** Account Id */
            account_id?: string | null;
            /**
             * Is Active
             * @default true
             */
            is_active: boolean;
            /** Deactivated At */
            deactivated_at?: string | null;
            /** Deactivated By User Id */
            deactivated_by_user_id?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}